Better command finding
Why:

The validate script only used to work with a globally-installed version of ESLint whose path was passed in through an environment variable. This is pretty inflexible. What if the project being validated wants to use a different version?

This change addresses the need by:

Instead of relying solely on a global eslint available through configuration to the script, the script first looks to see if there's a locally installed version of eslint and prefers that instead.